feat: add facility to persist game state
Leverage the vanilla localStorage JS and Recoil snapshot APIs to persist and save current game state of the board and the current player, within the active user browser.

The state of the game will persist / save only after the user places each connect piece on the board, and no other time.